refactor: auth 모듈 세분화, 패키지 조정 및 코드 정리
- auth 모듈을 adapter-in, adapter-out, application으로 세분화
- 로그인과 회원가입은 순수한 코틀린 객체로 구별하여 security 의존성을 낮춤
