+++ prestudy/README.md
@@ -4,7 +4,6 @@
 ### Architecture
 - 기본적으로 클린 아키텍처 적용
 - 모듈 설명
-  - `auth` : 인증/인가 관련 모듈, Spring Security 의존성이 높아 레이어드 아키텍처 적용
   - `bootstrap` : 스프링 애플리케이션 구동 모듈
   - `common` : 도메인 모듈에 공통적으로 사용하는 클래스를 모은 모듈
   - `domain` : 각 도메인에 해당하는 모듈
@@ -17,6 +16,8 @@
 - Spring Boot 3.3.2
   - Spring Web
   - Spring Data JPA
+  - Spring Security
+- JWT
 
 ### UseCase
 - 각 모듈별 http 파일 참조 (.http 파일 검색)
