--- prestudy/auth/build.gradle.kts
@@ -1,16 +1,0 @@
-plugins {
-    kotlin("plugin.jpa")
-}
-
-dependencies {
-    implementation(project(":common"))
-    implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-    implementation("org.springframework.boot:spring-boot-starter-security")
-    implementation("org.springframework.boot:spring-boot-starter-validation")
-    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.17.2")
-    runtimeOnly("com.h2database:h2:2.3.230") // runtimeOnly 만 제대로 작동함
-    implementation("io.jsonwebtoken:jjwt-api:0.12.6")
-    implementation("io.jsonwebtoken:jjwt-impl:0.12.6")
-    implementation("io.jsonwebtoken:jjwt-jackson:0.12.6")
-}
