--- prestudy/auth/src/main/kotlin/y/june/prestudy/auth/model/Member.kt
@@ -1,25 +1,0 @@
-package y.june.prestudy.auth.model
-
-import org.springframework.security.core.GrantedAuthority
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.userdetails.UserDetails
-
-class Member(
-    @JvmField
-    val username: String,
-    @JvmField
-    val password: String,
-    val role: Role,
-) : UserDetails {
-    override fun getAuthorities(): MutableCollection<out GrantedAuthority> {
-        return mutableListOf(SimpleGrantedAuthority(role.name))
-    }
-
-    override fun getPassword(): String {
-        return password
-    }
-
-    override fun getUsername(): String {
-        return username
-    }
-}
