--- prestudy/auth/src/main/kotlin/y/june/prestudy/auth/persistence/AuthTransactional.kt
@@ -1,13 +1,0 @@
-package y.june.prestudy.auth.persistence
-
-import org.springframework.transaction.annotation.Transactional
-import java.lang.annotation.Inherited
-
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
-@Retention
-@Inherited
-@Transactional(
-    transactionManager = AUTH_TRANSACTION_MANAGER,
-    rollbackFor = [Throwable::class]
-)
-annotation class AuthTransactional
