--- prestudy/auth/src/main/resources/application-auth.yaml
@@ -1,14 +1,0 @@
-spring:
-  datasource:
-    auth:
-      jdbc-url: jdbc:h2:mem:prestudy
-      username: sa
-      password:
-      driver-class-name: org.h2.Driver
-  # 이외 DB 설정은 JpaConfig 코드 참조
-  h2:
-    console:
-      enabled: true
-jwt:
-  duration: 3600 # (1시간)
-  secret: sldkfjwioejfskldjfopqfjopqpwr9fu90qru1029enfcv9032uq39rufq3pr903f9ui0dsufqvsxiodvcjko123knro3irf90
