+++ prestudy/domain/auth/adapter-in/build.gradle.kts
@@ -1,0 +1,5 @@
+dependencies {
+    implementation(project(":common"))
+    implementation(project(":domain:auth:auth-application"))
+    implementation("org.springframework.boot:spring-boot-starter-web")
+}
