+++ prestudy/domain/auth/adapter-out/src/main/resources/application-auth-adapter-out.yaml
@@ -1,0 +1,11 @@
+spring:
+  datasource:
+    auth:
+      jdbc-url: jdbc:h2:mem:prestudy
+      username: sa
+      password:
+      driver-class-name: org.h2.Driver
+  # 이외 DB 설정은 JpaConfig 코드 참조
+  h2:
+    console:
+      enabled: true
