+++ prestudy/domain/auth/application/src/main/kotlin/y/june/prestudy/auth/model/Member.kt
@@ -1,0 +1,8 @@
+package y.june.prestudy.auth.model
+
+class Member(
+    val id: Long?,
+    val username: String,
+    val password: String,
+    val role: Role,
+)
