+++ prestudy/domain/auth/application/src/main/kotlin/y/june/prestudy/auth/port/in/LoginUseCase.kt
@@ -1,0 +1,10 @@
+package y.june.prestudy.auth.port.`in`
+
+interface LoginUseCase {
+    fun login(command: LoginCommand): String
+}
+
+data class LoginCommand(
+    val username: String,
+    val password: String,
+)
