+++ prestudy/domain/auth/application/src/main/kotlin/y/june/prestudy/auth/port/in/SignUpUseCase.kt
@@ -1,0 +1,13 @@
+package y.june.prestudy.auth.port.`in`
+
+import y.june.prestudy.auth.model.Role
+
+interface SignUpUseCase {
+    fun singUp(command: SignUpCommand)
+}
+
+data class SignUpCommand(
+    val username: String,
+    val password: String,
+    val role: Role,
+)
