+++ prestudy/domain/auth/application/src/main/resources/application-auth-application.yaml
@@ -1,0 +1,3 @@
+jwt:
+  duration: 3600 # (1시간)
+  secret: sldkfjwioejfskldjfopqfjopqpwr9fu90qru1029enfcv9032uq39rufq3pr903f9ui0dsufqvsxiodvcjko123knro3irf90
